OrderDAO

+++ src/main/java/shop/User.java
@@ -3,7 +3,7 @@ package shop;
 
 public class User {
 
-	long id;
+	int id;
 	String name;
 	String surname;
 	String login;
@@ -17,17 +17,17 @@ public class User {
 		this.password = password;
 	}
 	
-	public User(String login, String password, long id) {
+	public User(String login, String password, int id) {
 		super();
 		this.login = login;
 		this.password = password;
 		this.id = id;
 	}
 	
-	public long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getName() {
